Initial demo of jQuery Tools issue
Problem is that the entire page is loaded with Ajax and then dumped in
the DOM when using jQuery Tools Overlay with external page load. This
includes the content of both the body AND the header. If the header has
scripts included then these script tags are both included in the DOM
append and loaded and interpreted. The script error can be seen as a
result.

Because this is supposed to be a progressive enhancement, the external
page requires a full HTML defintion, including header tags. The load()
method should only take the body section of the external page before
adding the content to the DOM, or there needs to be a method that only
loads the body.